move is_network_accessible helper to base.py
